fix: address pr review findings
