+++ crates/rginx-http/src/request_target/tests.rs
@@ -1,0 +1,17 @@
+use http::Uri;
+
+use super::{normalize_request_path, normalize_request_target};
+
+#[test]
+fn normalize_request_path_preserves_asterisk_form() {
+    assert_eq!(normalize_request_path("*"), "*");
+}
+
+#[test]
+fn normalize_request_target_preserves_asterisk_path_and_query_shape() {
+    let uri: Uri = "*".parse().expect("asterisk-form URI should parse");
+    let normalized = normalize_request_target(&uri);
+
+    assert_eq!(normalized.path, "*");
+    assert_eq!(normalized.path_and_query, "*");
+}
